+++ MinecraftJars.Plugin/MinecraftJars.Plugin.Purpur/Model/PurpurVersion.cs
@@ -6,7 +6,8 @@ namespace MinecraftJars.Plugin.Purpur.Model;
 
 public record PurpurVersion(
     IProject Project,
-    string Version) : IVersion
+    string Version,
+    bool IsSnapShot = false) : IVersion
 {
     public Task<IDownload> GetDownload(DownloadOptions? options = null, CancellationToken cancellationToken = default!) => 
         PurpurVersionFactory.GetDownload(options ?? new DownloadOptions(), this, cancellationToken);