+++ MinecraftJars.Plugin/MinecraftJars.Plugin.Spigot/Model/SpigotVersion.cs
@@ -7,7 +7,8 @@ namespace MinecraftJars.Plugin.Spigot.Model;
 public record SpigotVersion(
     IProject Project,
     string Version,
-    bool RequiresLocalBuild = false) : IVersion
+    bool RequiresLocalBuild = false,
+    bool IsSnapShot = false) : IVersion
 {
     internal DateTime? ReleaseTime { get; init; }
     internal string DetailUrl { get; init; } = string.Empty;